my god this is horrible wtf

+++ app/Types.hs
@@ -6,6 +6,7 @@ import Data.Bits
 import Control.Conditional (cond)
 import AST
 import Data.Char (ord, chr)
+import Data.Int (Int64)
 
 immShift = 3
 immMask = 0b111
@@ -21,28 +22,28 @@ typeInt = 0b0000
 typeChar = 0b01000
 maskChar = 0b11111
 
-valueToBits :: Expr -> Int
+valueToBits :: Expr -> Int64
 valueToBits e = case e of
-    Int i -> shiftL i intShift
+    Int i -> shiftL (fromIntegral i) intShift
     Bool b -> (if b then 0b00011000 else 0b00111000)
     Eof -> 0b01011000
-    Char c -> typeChar .|. shiftL (ord c) charShift 
+    Char c -> fromIntegral typeChar .|. shiftL (fromIntegral (ord c)) charShift
     _ -> -1
 
-bitsToValue :: Int -> Expr
+bitsToValue :: Int64 -> Expr
 
 bitsToValue b = cond [(valueToBits (Bool True) == b, Bool True )
                   , (valueToBits (Bool False) == b, Bool False )
                   , (valueToBits Eof == b, Eof )
-                  , (bitsIsInt b, Int (shiftR b intShift))
-                  , (bitsIsChar b, Char (chr (shiftR b charShift)))
+                  , (bitsIsInt b, Int (shiftR (fromIntegral b) intShift))
+                  , (bitsIsChar b, Char (chr (shiftR (fromIntegral b) charShift)))
                   , (otherwise , Eof)]
 
-bitsIsInt :: Int -> Bool
+bitsIsInt :: Int64 -> Bool
 bitsIsInt b = typeInt == (b .&. maskInt)
 
-bitsIsChar :: Int -> Bool
+bitsIsChar :: Int64 -> Bool
 bitsIsChar b = typeChar == (b .&. maskChar)
 
-bitsIsImm :: Int -> Bool
+bitsIsImm :: Int64 -> Bool
 bitsIsImm b = 0 == (b .&. immMask)
